tests: Move tests from test_utils.py to test_file_utils.py
